pr14rev1

+++ controllers/cards.js
@@ -1,45 +1,54 @@
 const Card = require('../models/card');
+const NotFoundError = require('../errors/not-found-err');
+const BadRequestError = require('../errors/bad-request-err');
+const UnauthorizedError = require('../errors/unauthorized-err');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({}).then((data) => {
     if (data.length >= 1) {
       res.send({ cards: data });
     } else {
       res.status(200).send({ message: 'Карточки не найдены' });
     }
-  }).catch(() => {
-    res.status(500).send({ message: 'Ошибка по умолчанию' });
+  }).catch((err) => {
+    next(err);
   });
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id }).then((data) => {
     res.send({ card: data });
   }).catch((err) => {
     if (err.name === 'ValidationError') {
-      res.status(400).send({ message: 'Переданы некорректные данные при создании карточки' });
+      throw new BadRequestError('Переданы некорректные данные при создании карточки');
     } else {
-      res.status(500).send({ message: 'Ошибка по умолчанию' });
+      next(err);
     }
   });
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId).then((data) => {
-    if (data) {
-      res.send({ card: data });
-    } else { res.status(404).send({ message: 'Карточка с указанным _id не найдена' }); }
-  }).catch((err) => {
-    if (err.name === 'CastError') {
-      res.status(400).send({ message: 'Переданы некорректные данные при удалении карточки' });
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId).then((card) => {
+    if (req.user._id === card.owner.toString()) {
+      Card.findByIdAndRemove(req.params.cardId).then((data) => {
+        if (data) {
+          res.status(200).send({ card: data });
+        } else { throw new NotFoundError('Карточка с указанным _id не найдена'); }
+      }).catch((err) => {
+        if (err.name === 'CastError') {
+          throw new BadRequestError('Переданы некорректные данные при удалении карточки');
+        } else {
+          next(err);
+        }
+      });
     } else {
-      res.status(500).send({ message: 'Ошибка по умолчанию' });
+      throw new UnauthorizedError('Вы не можете удалять чужие карточки');
     }
   });
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -47,17 +56,17 @@ module.exports.likeCard = (req, res) => {
   ).then((data) => {
     if (data) {
       res.send({ card: data });
-    } else { res.status(404).send({ message: 'Передан несуществующий _id карточки' }); }
+    } else { throw new NotFoundError('Передан несуществующий _id карточки'); }
   }).catch((err) => {
     if (err.name === 'CastError') {
-      res.status(400).send({ message: 'Переданы некорректные данные для постановки лайка' });
+      throw new BadRequestError('Переданы некорректные данные для постановки лайка');
     } else {
-      res.status(500).send({ message: 'Ошибка по умолчанию' });
+      next(err);
     }
   });
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -65,12 +74,12 @@ module.exports.dislikeCard = (req, res) => {
   ).then((data) => {
     if (data) {
       res.send({ card: data });
-    } else { res.status(404).send({ message: 'Передан несуществующий _id карточки' }); }
+    } else { throw new NotFoundError('Передан несуществующий _id карточки'); }
   }).catch((err) => {
     if (err.name === 'CastError') {
-      res.status(400).send({ message: 'Переданы некорректные данные для снятии лайка' });
+      throw new BadRequestError('Переданы некорректные данные для снятии лайка');
     } else {
-      res.status(500).send({ message: 'Ошибка по умолчанию' });
+      next(err);
     }
   });
 };
